Fix failing test cases when JFR exists
