+++ src/main/java/jdk/jfr/FlightRecorder.java
@@ -27,7 +27,6 @@ package jdk.jfr;
 
 import java.util.Collections;
 import java.util.List;
-import java.util.Objects;
 
 import jdk.jfr.internal.Warnings;
