+++ src/test/java/jfr/compat/test/util/TestToolkit.java
@@ -1,0 +1,102 @@
+/*
+ * Copyright (c) 2016, 2018, Oracle and/or its affiliates. All rights reserved.
+ * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
+ *
+ * This code is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License version 2 only, as
+ * published by the Free Software Foundation.  Oracle designates this
+ * particular file as subject to the "Classpath" exception as provided
+ * by Oracle in the LICENSE file that accompanied this code.
+ *
+ * This code is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
+ * version 2 for more details (a copy is included in the LICENSE file that
+ * accompanied this code).
+ *
+ * You should have received a copy of the GNU General Public License version
+ * 2 along with this work; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+ *
+ * Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
+ * or visit www.oracle.com if you need additional information or have any
+ * questions.
+ */
+package jfr.compat.test.util;
+
+import java.io.Closeable;
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
+
+public class TestToolkit {
+
+	public static File materialize(String dir, String file, Class<?> clazz) {
+		File matDir;
+		try {
+			matDir = materialize(clazz, dir, file);
+			File matFile = new File(matDir, file);
+			return matFile;
+		} catch (IOException e) {
+			e.printStackTrace();
+		}
+		return null;
+	}
+	private static File materialize(Class<?> clazz, String directoryName, String fileName) throws IOException {
+		File directory = File.createTempFile(directoryName, ".dir");
+		materialize(clazz, fileName, directory);
+		return directory;
+	}
+
+	private static void materialize(Class<?> clazz, String fileName, File directory)
+			throws IOException {
+		if (fileName == null) {
+			throw new IOException("Must specify file name to materialize");
+		}
+		if (!directory.delete()) {
+			throw new IOException("Could not delete directory: " + directory.getAbsolutePath());
+		}
+		if (!directory.mkdirs()) {
+			throw new IOException("Could not create directory: " + directory.getAbsolutePath());
+		}
+		InputStream in = null;
+		try {
+			in = clazz.getResourceAsStream(fileName);
+			if (in != null) {
+				FileOutputStream os = null;
+				try {
+					File file = new File(directory, fileName);
+					os = new FileOutputStream(file);
+					copy(in, os);
+					os.close();
+				} finally {
+					closeSilently(os);
+				}
+			}
+		} finally {
+			closeSilently(in);
+		}
+	}
+
+	private static void copy(InputStream is, OutputStream os) throws IOException {
+		int length;
+		byte[] buffer = new byte[1024];
+		while ((length = is.read(buffer)) > 0) {
+			os.write(buffer, 0, length);
+		}
+		is.close();
+	}
+
+	private static void closeSilently(Closeable closeable) {
+		if (closeable != null) {
+			try {
+				closeable.close();
+			} catch (IOException e) {
+				// keep your mouth shut
+			}
+		}
+	}
+
+}
